+++ packages/react-noop-renderer/index.js
@@ -1,4 +1,3 @@
 'use strict';
 
-// TODO: Move the source into packages
-module.exports = require('ReactNoopEntry');
+module.exports = require('./src/ReactNoopEntry');
