+++ packages/react-noop-renderer/src/ReactNoopEntry.js
